consolidate exception classes
