+++ taskbadger/sdk.py
@@ -4,8 +4,7 @@ from typing import List
 
 from _contextvars import ContextVar
 
-from taskbadger.errors import ServerError, Unauthorized, UnexpectedStatus
-from taskbadger.exceptions import ConfigurationError
+from taskbadger.exceptions import ConfigurationError, ServerError, Unauthorized, UnexpectedStatus
 from taskbadger.integrations import Action
 from taskbadger.internal import AuthenticatedClient
 from taskbadger.internal.api.task_endpoints import task_create, task_get, task_partial_update
